Logic work
Deciding the data structures needed. I went with an array to hold Game Objects then set the pieces to prefabs and pre-set the teams on each piece. the tiles will also probably be an array of Game Objects but 2d. Set up some of the properties of the classes used in the game.