Added Pieces
I added the pieces to the board. They were also assigned into the array within the board script. The tiles array is now a 2d array that automatically finds the correct tiles for each index on start.

I can begin to work on the logic for the game now that set up is complete.