Castling
I added the castling rule to allow the players to castle.

This was done using a new Move Plate function and taking another reference piece when the castle bool is true. This allows me to move the position of the king and rook it castled with. Then by checking if the rook or the king had a previous move I update the new positions of the pieces.